Modify the current modal form to be a new form component and create a new form component for deleting a card
